run next commands by default
