+++ packages/blitz/src/cli/commands/next/index.ts
@@ -1,0 +1,5 @@
+import {dev} from "./dev"
+import {build} from "./build"
+import {start} from "./start"
+
+export {dev, build, start}
